+++ src/main/java/rs/ac/uns/ftn/BookingBaboon/controllers/accommodation_handling/AccommodationModificationController.java
@@ -5,6 +5,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import rs.ac.uns.ftn.BookingBaboon.domain.accommodation_handling.Accommodation;
 import rs.ac.uns.ftn.BookingBaboon.domain.accommodation_handling.AccommodationModification;
 import rs.ac.uns.ftn.BookingBaboon.dtos.accommodation_handling.accommodation_modification.AccommodationModificationCreateRequest;
 import rs.ac.uns.ftn.BookingBaboon.dtos.accommodation_handling.accommodation_modification.AccommodationModificationRequest;
@@ -13,7 +14,7 @@ import rs.ac.uns.ftn.BookingBaboon.services.accommodation_handling.interfaces.IA
 
 import java.util.Collection;
 import java.util.stream.Collectors;
-
+@CrossOrigin
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api/v1/accommodation-change-requests")
@@ -31,8 +32,13 @@ public class AccommodationModificationController {
 
     @GetMapping("/{id}")
     public ResponseEntity<AccommodationModificationResponse> get(@PathVariable Long id) {
-        AccommodationModificationResponse response = mapper.map(service.get(id), AccommodationModificationResponse.class);
-        return new ResponseEntity<>(response,  HttpStatus.OK) ;
+        AccommodationModificationResponse result = mapper.map(service.get(id), AccommodationModificationResponse.class);
+
+        if (result == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>(result,  HttpStatus.OK) ;
     }
 
     @PostMapping
@@ -44,26 +50,46 @@ public class AccommodationModificationController {
     @PutMapping
     public ResponseEntity<AccommodationModificationResponse> update(@RequestBody AccommodationModificationRequest request) {
         AccommodationModification result = service.update(mapper.map(request, AccommodationModification.class));
+
+        if (result == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         return new ResponseEntity<>(mapper.map(result, AccommodationModificationResponse.class), HttpStatus.OK);
     }
 
     @PutMapping("/approve/{id}")
     public ResponseEntity<AccommodationModificationResponse> approve(@PathVariable Long id) {
-        AccommodationModification request = service.get(id);
-        request.Approve();
-        return new ResponseEntity<>(mapper.map(request, AccommodationModificationResponse.class), HttpStatus.OK);
+        AccommodationModification accommodationModification = service.get(id);
+
+        if (accommodationModification == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        accommodationModification.Approve();
+        return new ResponseEntity<>(mapper.map(accommodationModification, AccommodationModificationResponse.class), HttpStatus.OK);
     }
 
     @PutMapping("/deny/{id}")
     public ResponseEntity<AccommodationModificationResponse> deny(@PathVariable Long id) {
-        AccommodationModification request = service.get(id);
-        request.Deny();
-        return new ResponseEntity<>(mapper.map(request, AccommodationModificationResponse.class), HttpStatus.OK);
+        AccommodationModification accommodationModification = service.get(id);
+
+        if (accommodationModification == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        accommodationModification.Deny();
+        return new ResponseEntity<>(mapper.map(accommodationModification, AccommodationModificationResponse.class), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> remove(@PathVariable Long id) {
-        service.remove(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        AccommodationModification accommodationModification = service.get(id);
+        if (accommodationModification != null) {
+            service.remove(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
